Moved ImgGetter and BitmapGetter outside of ImgListener
