+++ app/src/main/java/com/fourlastor/dante/MainActivity.java
@@ -9,7 +9,7 @@ import android.text.Spanned;
 import android.widget.TextView;
 
 import com.fourlastor.dante.html.FlavoredHtml;
-import com.fourlastor.dante.html.ImgListener;
+import com.fourlastor.dante.html.ImgGetter;
 import com.squareup.picasso.Picasso;
 
 import java.io.IOException;
@@ -35,7 +35,7 @@ public class MainActivity extends AppCompatActivity {
                 .style(Typeface.BOLD, "b", "strong")
                 .style(Typeface.ITALIC, "i", "em")
                 .bullet(15, "li")
-                .img(new ImgListener.BitmapImgGetter(getResources()) {
+                .img(new ImgGetter.BitmapImgGetter(getResources()) {
                     @Override
                     protected Bitmap getBitmap(String src) {
                         try {
